// src/Moryx.ControlSystem.WorkerSupport.Web/src/app/api/strict-http-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpResponse } from '@angular/common/http';

/**
 * Constrains the http response to not have the body defined as `T | null`, but `T` only.
 */
export type StrictHttpResponse<T> = HttpResponse<T> & {
  readonly body: T;
}